fix: исправить RequestPrivateService

+++ ewm-general-service/src/main/java/ru/practikum/ewm/general/model/dto/EventFullDto.java
@@ -16,7 +16,7 @@ public class EventFullDto {
     String title;
     LocationDto location;
     long views;
-    int participantLimit;
+    long participantLimit;
     boolean requestModeration;
     String publishedOn;
     EventState state;
